+++ server/Extensions/HangfireExtensions.cs
@@ -1,0 +1,22 @@
+using Hangfire;
+using Hangfire.PostgreSql;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Fortedle.Server.Extensions;
+
+public static class HangfireExtensions
+{
+    public static IServiceCollection AddHangfireServices(this IServiceCollection services, string connectionString)
+    {
+        // Configure Hangfire
+        services.AddHangfire(config => config
+            .SetDataCompatibilityLevel(CompatibilityLevel.Version_180)
+            .UseSimpleAssemblyNameTypeSerializer()
+            .UseRecommendedSerializerSettings()
+            .UsePostgreSqlStorage(options => options.UseNpgsqlConnection(connectionString)));
+
+        services.AddHangfireServer();
+
+        return services;
+    }
+}